Add asset removal button
